# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Project 2 API
  description: API documentation for CSE 341 Project 2
  version: "1.0.0"
servers:
  - url: https://cse341-project2-elpb.onrender.com
security:
  - BearerAuth: []
paths:
...
  /users/{id}:
    get:
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to retrieve
      responses:
        '200':
          description: User data retrieved successfully
        '404':
          description: User not found
        '500':
          description: Server error

    put:
      summary: Update a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: updateduser@example.com
                password:
                  type: string
                  example: newPassword123 (optional)
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Validation error
        '404':
          description: User not found or not updated
        '500':
          description: Server error during update

    delete:
      summary: Delete a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to delete
      responses:
        '200':
          description: User deleted successfully
        '404':
          description: User not found
        '500':
          description: Server error during delete
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...